+++ src/utils/constants.ts
@@ -1,0 +1,3 @@
+export const pathSepRegex = /[\\/]/;
+export const windowsInvalidName = /[<>:"|?*]/;
+export const isWindows = process.platform === 'win32';